Updated shared installer. Use as default as simplifies linking issues on Cori

+++ strumpack/STRUMPACK_installer_shared.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# This script aims to download and build all dependencies for the installation of STRUMPACK in an automated way.
+# Differences may exist between machines, and work on this is tested mostly using the NERSC Cori supercomputer.
+
 # Set urls for packages
 STRUMPACK=https://github.com/pghysels/STRUMPACK.git
 METIS=http://glaros.dtc.umn.edu/gkhome/fetch/sw/metis/metis-5.1.0.tar.gz
@@ -14,23 +17,24 @@ CMAKE=https://cmake.org/files/v3.9/
 OPENMPI=https://www.open-mpi.org/software/ompi/v3.0/downloads/openmpi-3.0.0.tar.gz
 
 # Create directory structure
-if [ ! -d ./downloads ]; then
-  mkdir downloads
+if [ ! -d ./strumpack_downloads ]; then
+  mkdir strumpack_downloads
 fi
-if [ ! -d ./deps ]; then
-  mkdir deps
+if [ ! -d ./strumpack_deps ]; then
+  mkdir strumpack_deps
 fi
-if [ ! -d ./builds ]; then
-  mkdir builds
+if [ ! -d ./strumpack_build ]; then
+  mkdir strumpack_build
 fi
 
 pushd . >/dev/null
-cd downloads
+cd strumpack_downloads
 
 # Setup NERSC environment for building
 if [ "$NERSC_HOST" == "cori" ]; then
   module swap PrgEnv-intel PrgEnv-gnu
   module swap gcc gcc/4.9.3
+  module load openmpi
   #module load python/2.7-anaconda
   #source activate ${1} ##Pass conda env name as argument
 fi
@@ -38,13 +42,13 @@ fi
 #################################################
 # Acquire STRUMPACK-sparse and all dependencies #
 #################################################
-if [ ! -d ../deps/STRUMPACK ]; then
+if [ ! -d ../strumpack_deps/STRUMPACK ]; then
   echo "Downloading STRUMPACK"
   git clone $STRUMPACK
-  cp -Rf STRUMPACK ../deps
+  cp -Rf STRUMPACK ../strumpack_deps
 else
   pushd . > /dev/null
-  cd ../deps/STRUMPACK;
+  cd ../strumpack_deps/STRUMPACK;
   git pull --rebase
   popd > /dev/null
 fi
@@ -53,20 +57,19 @@ fi
 if [ ! -e OPENMPI.tar.gz ]; then
   echo "Downloading OPENMPI"
   curl -L $OPENMPI -o OPENMPI.tar.gz
-  mkdir ../deps/OPENMPI
-  tar xvf OPENMPI.tar.gz -C ../deps/OPENMPI --strip-components 1 
+  mkdir ../strumpack_deps/OPENMPI
+  tar xvf OPENMPI.tar.gz -C ../strumpack_deps/OPENMPI --strip-components 1 
 fi
-
 
 # METIS v5.1.0 nested dissection
 if [ ! -e METIS.tar.gz ]; then
   echo "Downloading METIS"
   curl -L $METIS -o METIS.tar.gz
-  tar xvf METIS.tar.gz -C ../deps
+  tar xvf METIS.tar.gz -C ../strumpack_deps
   #Update library to 64 bit width
   if [ $(getconf LONG_BIT) == "64" ]; then
-    sed -i.bak 's/IDXTYPEWIDTH 32/IDXTYPEWIDTH 64/g' ../deps/metis-5.1.0/include/metis.h;
-    rm ../deps/metis-5.1.0/include/metis.h.bak
+    sed -i.bak 's/IDXTYPEWIDTH 32/IDXTYPEWIDTH 64/g' ../strumpack_deps/metis-5.1.0/include/metis.h;
+    rm ../strumpack_deps/metis-5.1.0/include/metis.h.bak
   fi
 fi
 
@@ -74,11 +77,11 @@ fi
 if [ ! -e PARMETIS.tar.gz ]; then
   echo "Downloading PARMETIS"
   curl -L $PARMETIS -o PARMETIS.tar.gz
-  tar xvf PARMETIS.tar.gz -C ../deps
+  tar xvf PARMETIS.tar.gz -C ../strumpack_deps
     #Update library to 64 bit width
   if [ $(getconf LONG_BIT) == "64" ]; then
-    sed -i.bak 's/IDXTYPEWIDTH 32/IDXTYPEWIDTH 64/g' ../deps/parmetis-4.0.3/metis/include/metis.h;
-    rm ../deps/parmetis-4.0.3/metis/include/metis.h.bak
+    sed -i.bak 's/IDXTYPEWIDTH 32/IDXTYPEWIDTH 64/g' ../strumpack_deps/parmetis-4.0.3/metis/include/metis.h;
+    rm ../strumpack_deps/parmetis-4.0.3/metis/include/metis.h.bak
   fi
 fi
 
@@ -86,7 +89,7 @@ fi
 if [ ! -e SCOTCH.tar.gz ]; then
   echo "Downloading SCOTCH"
   curl -L $SCOTCH -o SCOTCH.tar.gz
-  tar xvf SCOTCH.tar.gz -C ../deps
+  tar xvf SCOTCH.tar.gz -C ../strumpack_deps
 fi
 
 # TCMALLOC: Faster malloc/new; Can potentially also use tbbmalloc 
@@ -95,48 +98,48 @@ fi
 if [ ! -e TCMALLOC.tar.gz ]; then
   echo "Downloading TCMALLOC"
   curl -L $TCMALLOC -o TCMALLOC.tar.gz
-  tar xvf TCMALLOC.tar.gz -C ../deps
+  tar xvf TCMALLOC.tar.gz -C ../strumpack_deps
 fi
 
 # OPENBLAS: Open source BLAS library
-if [ ! -e OPENBLAS.tar.gz ]; then
+if [ ! -e OPENBLAS.tar.gz ] && [ "$NERSC_HOST" != "cori" ]; then
   echo "Downloading OPENBLAS"
   curl -L $OPENBLAS -o OPENBLAS.tar.gz
-  tar xvf OPENBLAS.tar.gz -C ../deps
+  tar xvf OPENBLAS.tar.gz -C ../strumpack_deps
 fi
 
 # CMAKE
-if [ ! -e CMAKE.tar.gz ]; then
+if [ ! -e CMAKE.tar.gz ] && [ "$NERSC_HOST" != "cori" ]; then
   echo "Downloading CMAKE"
   if [[ "$OSTYPE" == "linux"* ]]; then
     curl -L $CMAKE/cmake-3.9.1-Linux-x86_64.tar.gz -o CMAKE.tar.gz
-    export CMAKE_DIR=$PWD/../deps/cmake-3.9.1-Linux-x86_64/
+    export CMAKE_DIR=$PWD/../strumpack_deps/cmake-3.9.1-Linux-x86_64/
   elif [[ "$OSTYPE" == "darwin"* ]];then 
     curl -L $CMAKE/cmake-3.9.1-Darwin-x86_64.tar.gz -o CMAKE.tar.gz
-    export CMAKE_DIR=$PWD/../deps/cmake-3.9.1-Darwin-x86_64/CMake.app/Contents/
+    export CMAKE_DIR=$PWD/../strumpack_deps/cmake-3.9.1-Darwin-x86_64/CMake.app/Contents/
   fi
-  tar xvf CMAKE.tar.gz -C ../deps
-  cp -Rv $CMAKE_DIR/* ../builds
+  tar xvf CMAKE.tar.gz -C ../strumpack_deps
+  cp -Rv $CMAKE_DIR/* ../strumpack_build
 fi
 
 # SCALAPACK
 if [ ! -e SCALAPACK.tar.gz ]; then
   echo "Downloading SCALAPACK"
   curl -L $SCALAPACK -o SCALAPACK.tar.gz
-  tar xvf SCALAPACK.tar.gz -C ../deps
+  tar xvf SCALAPACK.tar.gz -C ../strumpack_deps
 fi
 
 popd > /dev/null
 pushd . > /dev/null;
 
 # Set the local bin directory on path for cmake, etc, as well as include paths
-export INSTALL_DIR=$PWD/builds
+export INSTALL_DIR=$PWD/strumpack_build
 export prefix=$INSTALL_DIR
-export PATH=$PATH:$PWD/builds/bin
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$PWD/builds/lib:$PWD/builds/lib64
+export PATH=$PATH:$PWD/strumpack_build/bin
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$PWD/strumpack_build/lib:$PWD/strumpack_build/lib64
 
-LIB_DIR="-L$PWD/builds/lib -L$PWD/builds/lib64 -L$CONDA_PREFIX/lib "
-INC_DIR="-I$PWD/builds/include -I$CONDA_PREFIX/include "
+LIB_DIR="-L$PWD/strumpack_build/lib -L$PWD/strumpack_build/lib64 -L$CONDA_PREFIX/lib "
+INC_DIR="-I$PWD/strumpack_build/include -I$CONDA_PREFIX/include "
 
 #Append Cray include and libs to build flags where necessary
 if [ "$NERSC_HOST" == "cori" ]; then
@@ -150,21 +153,20 @@ fi
 
 #Get core count/2
 let proc=$(getconf _NPROCESSORS_ONLN)/2
-cd deps
+cd strumpack_deps
 
 # Build all dependencies and dependency dependencies
 
 #OPENMPI: Optional if to be included from modules/conda
-if [ "$NERSC_HOST" != "cori" ]; then
-  cd ./OPENMPI
-  ./configure --prefix=$INSTALL_DIR --enable-mpi-thread-multiple
-  make -j$(echo ${proc}) && make install
-  cd ..
-else
-  alias mpicc='cc -fPIC'
-  alias mpicxx='CC -fPIC'
-fi
-if 0==1; then
+#if [ "$NERSC_HOST" != "cori" ]; then
+#  cd ./OPENMPI
+#  ./configure --prefix=$INSTALL_DIR --enable-mpi-thread-multiple --with-verbs=no
+#  make -j$(echo ${proc}) && make install
+#  cd ..
+#else
+#  alias mpicc='cc -fPIC'
+#  alias mpicxx='CC -fPIC'
+#fi
 cd ./metis-5.1.0;
 if [ "$NERSC_HOST" != "cori" ]; then
   make config cc='mpicc' shared=1 prefix=$INSTALL_DIR
@@ -183,22 +185,33 @@ else
 fi
 make -j$(echo ${proc}) && make install
 cd ..
+
 # Need to choose correct Makefile from those given; prefix needs to be passed as env variable; Path to include dir needed for MPI
 # Expects mpi.h to be in /usr/include; not ideal;
 cd ./scotch_6.0.4/src
 cp ./Make.inc/Makefile.inc.x86-64_pc_linux2.shlib ./Makefile.inc
-sed -i.bak 's@-O3@'"-O3 $INC_DIR"'@' ./Makefile.inc #Add CFLAGS env variable into compile path for mpi headers
-sed -i.bak 's/-DSCOTCH_PTHREAD//' ./Makefile.inc #Disable scotch pthreads as causes issues with MPI
-sed -i.bak 's/-pthread//' ./Makefile.inc #Disable scotch pthreads as causes issues with MPI
+#sed -i.bak 's@-O3@'"-O3 $INC_DIR"'@' ./Makefile.inc #Add CFLAGS env variable into compile path for mpi headers
+sed -i.bak 's@-O3@'"-O3 "'@' ./Makefile.inc #Add CFLAGS env variable into compile path for mpi headers
 if [ "$NERSC_HOST" == "cori" ]; then
   sed -i.bak 's/gcc/cc/' ./Makefile.inc #Change default compiler
   sed -i.bak 's/mpicc/cc/' ./Makefile.inc #Change default compiler for mpi
 fi
 rm ./Makefile.inc.bak
-make scotch -j$(echo ${proc}) && make ptscotch -j$(echo ${proc}) && make install
+#Patch scotch makefile to exclude everything except the shared libraries
+echo '
+install_strumpack	:	required    $(includedir)   $(libdir)   $(mandir)/man1 
+						-$(CP) -f ../include/*scotch*.h $(includedir) 
+						-$(CP) -f ../lib/*scotch*$(LIB) $(libdir) 
+						-$(CP) -Rf ../man/* $(mandir) 
+scotch_strumpack	:	required 
+						(cd libscotch ; $(MAKE) VERSION=$(VERSION) RELEASE=$(RELEASE) PATCHLEVEL=$(PATCHLEVEL) scotch && \
+										$(MAKE) VERSION=$(VERSION) RELEASE=$(RELEASE) PATCHLEVEL=$(PATCHLEVEL) install && \
+										$(MAKE) VERSION=$(VERSION) RELEASE=$(RELEASE) PATCHLEVEL=$(PATCHLEVEL) ptscotch && \
+										$(MAKE) VERSION=$(VERSION) RELEASE=$(RELEASE) PATCHLEVEL=$(PATCHLEVEL) ptinstall )
+' >> ./Makefile
+make scotch_strumpack -j$(echo ${proc}) && make prefix=$INSTALL_DIR install_strumpack -j$(echo ${proc}) 
 cd ../..
 
-fi
 # Install OpenBLAS; MKL might be a good option too
 if [ "$NERSC_HOST" != "cori" ]; then
   cd ./OpenBLAS-0.2.20
@@ -212,6 +225,7 @@ if [ "$NERSC_HOST" != "cori" ]; then
   make -j$(echo ${proc}) && make install
   cd ..
 fi
+echo "Finished"
 
 # STRUMPACK build
 # Parameterise for Cori login/Haswell, KNL, or non-Cori linux system
@@ -221,8 +235,8 @@ mkdir build
 cd build
 if [ "$NERSC_HOST" != "cori" ]; then #Not on a Cori node; use the locally built packages for all dependencies
   echo "Standard installation"
-  $INSTALL_DIR/bin/cmake ../ -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR \
-  -DCMAKE_CXX_COMPILER=$INSTALL_DIR/bin/mpic++ -DCMAKE_C_COMPILER=$INSTALL_DIR/bin/mpicc \
+  $INSTALL_DIR/bin/cmake ../  -DBUILD_SHARED_LIBS:BOOL=ON -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR \
+ -DCMAKE_CXX_COMPILER=$INSTALL_DIR/bin/mpic++ -DCMAKE_C_COMPILER=$INSTALL_DIR/bin/mpicc \
   -DCMAKE_Fortran_COMPILER=$INSTALL_DIR/bin/mpifort \
   -DSCALAPACK_LIBRARIES="$INSTALL_DIR/lib/libscalapack.a" \
   -DMETIS_INCLUDES=$INSTALL_DIR/include -DMETIS_LIBRARIES=$INSTALL_DIR/lib/libmetis.a \
@@ -238,7 +252,7 @@ elif [[ $(getconf _NPROCESSORS_ONLN) -ne 272 ]]; then #If not 272 cores, then we
   -DSCOTCH_LIBRARIES="$INSTALL_DIR/lib/libscotch.so;$INSTALL_DIR/lib/libscotcherr.so;$INSTALL_DIR/lib/libptscotch.so;$INSTALL_DIR/lib/libptscotcherr.so"
 else #Otherwise on a KNL node; Need to adjust so that the libraries used as KNL specific
   echo "KNL installation"
-  cmake ../ -D_GLIBCXX_USE_CXX11_ABI=0 -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR \
+  cmake ../ -D_GLIBCXX_USE_CXX11_ABI=0  -DBUILD_SHARED_LIBS:BOOL=ON -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR \
    -DCMAKE_CXX_COMPILER=CC -DCMAKE_C_COMPILER=cc -DCMAKE_Fortran_COMPILER=ftn \
    -DCMAKE_EXE_LINKER_FLAGS="-dynamic" -DMETIS_INCLUDES=$INSTALL_DIR/include \
    -DMETIS_LIBRARIES=$INSTALL_DIR/lib/libmetis.a -DPARMETIS_INCLUDES=$INSTALL_DIR/include \
@@ -247,3 +261,15 @@ else #Otherwise on a KNL node; Need to adjust so that the libraries used as KNL 
 fi
 make -j ${proc} && make install
 popd > /dev/null
+
+cd ./build
+cp ./dispatcher_include_template.sh dispatcher_include_strumpack.sh
+str='if [ -n "$LD_LIBRARY_PATH" ]; then \
+  LD_LIBRARY_PATH="$LIBTBX_BUILD/../strumpack_build/lib:$LD_LIBRARY_PATH" \
+  export LD_LIBRARY_PATH \
+else \
+  LD_LIBRARY_PATH="$LIBTBX_BUILD/../strumpack_build/lib" \
+  export LD_LIBRARY_PATH \
+ fi'
+sed -i "3 a ${str}" dispatcher_include_strumpack.sh
+libtbx.refresh
